Split up the importers so ISC importer has a parallel home to JSON importer.
